# src/main/java/com/example/bankcards/dto/user/request/CardTransactionRequestDto.java
package com.example.bankcards.dto.user.request;

import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.NotNull;
import java.math.BigDecimal;
import java.util.UUID;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CardTransactionRequestDto {

    @NotNull(message = "Output card reference is required")
    private UUID fromCardRef;

    @NotNull(message = "Input card reference is required")
    private UUID toCardRef;

    @NotNull(message = "Transfer amount is required")
    @DecimalMin(value = "0.01", message = "Amount must be greater than zero")
    private BigDecimal amount;
}